requirements.txt
flask==2.2.3
google-generativeai==0.3.1
python-dotenv==1.0.0
Werkzeug==2.2.3 